+++ ts_control/src/serve.rs
@@ -683,6 +683,7 @@ mod tests {
             id: 1,
             stable_id: StableId("n1".to_string()),
             hostname: "host".to_string(),
+            user_id: 0,
             tailnet: Some("tail1.ts.net".to_string()),
             tags: vec![],
             tailnet_address: TailnetAddress {